PAE-1336: move is-report-complete to application layer
