+++ src/reports/application/is-report-complete.js
@@ -4,10 +4,10 @@ import {
   recyclingActivitySchema
 } from '#reports/repository/schema.js'
 import Joi from 'joi'
-import { OPERATOR_CATEGORY } from './operator-category.js'
+import { OPERATOR_CATEGORY } from '#reports/domain/operator-category.js'
 
 /**
- * @import { OperatorCategory } from './operator-category.js'
+ * @import { OperatorCategory } from '#reports/domain/operator-category.js'
  * @import { Report } from '#reports/repository/port.js'
  */
