+++ src/reports/routes/status.js
@@ -3,8 +3,8 @@ import Joi from 'joi'
 import { StatusCodes } from 'http-status-codes'
 
 import { auditReportStatusTransition } from '#reports/application/audit.js'
+import { isReportComplete } from '#reports/application/is-report-complete.js'
 import { fetchCurrentReport } from '#reports/application/report-service.js'
-import { isReportComplete } from '#reports/domain/is-report-complete.js'
 import { getOperatorCategory } from '#reports/domain/operator-category.js'
 import { REPORT_STATUS } from '#reports/domain/report-status.js'
 import { isValidReportTransition } from '#reports/domain/report-transitions.js'
